+++ source/benoit/benoit/launcher/print_message.rs
@@ -21,39 +21,16 @@
 	If not, see <https://www.gnu.org/licenses/>.
 */
 
-use crate::benoit::complex::Complex;
-use crate::benoit::fractal::Fractal;
-use crate::benoit::image::ImageFormat;
-use crate::benoit::palette::Palette;
+use crate::benoit::VERSION;
+use crate::benoit::launcher::Launcher;
 
-extern crate rug;
-
-use rug::Float;
-
-pub mod animate;
-pub mod configure;
-pub mod dump_frame;
-pub mod run;
-pub mod still;
-
-pub struct Script {
-	// Configuration:
-	fractal:  Fractal,
-
-	canvas_width:  u32,
-	canvas_height: u32,
-	frame_start:   u32,
-	frame_stop:    u32,
-
-	centre: Complex,
-	extra:  Complex,
-	zoom:   Float,
-
-	max_iter_count: u32,
-
-	palette:      Palette,
-	colour_range: f32,
-
-	dump_path:    String,
-	image_format: ImageFormat,
+impl Launcher {
+	pub(super) fn print_message() {
+		println!();
+		println!(" \u{1B}[1mBENO\u{CE}T\u{1B}[0m {:X}.{:X}.{:X}", VERSION.0, VERSION.1, VERSION.2);
+		println!(" Copyright 2021, 2023 \u{1B}[1mGabriel Bj\u{F8}rnager Jensen\u{1B}[0m.");
+		println!();
+		println!(" \u{1B}[3mLe p\u{E8}re cogita et c'est pourquoi il fut.\u{1B}[0m");
+		println!();
+	}
 }