+++ source/benoit/main.rs
@@ -23,59 +23,13 @@
 
 mod benoit;
 
-use crate::benoit::VERSION;
-use crate::benoit::app::App;
-use crate::benoit::configuration::Configuration;
-use crate::benoit::palette::fill_palettes;
-use crate::benoit::script::Script;
+use crate::benoit::launcher::Launcher;
 
-extern crate rayon;
-
-use rayon::ThreadPoolBuilder;
-use std::env::args;
 use std::process::exit;
-use std::thread::available_parallelism;
 
 fn main() {
-	println!();
-	println!(" \u{1B}[1mBENO\u{CE}T\u{1B}[0m {:X}.{:X}.{:X}", VERSION.0, VERSION.1, VERSION.2);
-	println!(" Copyright 2021, 2023 Gabriel Bj\u{F8}rnager Jensen.");
-	println!();
-	println!(" \u{1B}[3mLe p\u{E8}re cogita et c'est pourquoi il fut.\u{1B}[0m");
-	println!();
-
-	fill_palettes();
-
-	let mut arguments = args();
-
-	let (mut configuration, interative) = match arguments.nth(0x1) {
-		Some(path) => (Configuration::load(path.as_str()), false),
-		_          => (Configuration::default(),           true),
-	};
-
-	configuration.thread_count = if configuration.thread_count == 0x0 {
-		match available_parallelism() {
-			Ok(ammount) => ammount.get() as u32,
-			_           => 0x2, // We assume at least two threads.
-		}
-	} else {
-		configuration.thread_count
-	};
-
-	eprintln!("using {} threads", configuration.thread_count);
-	ThreadPoolBuilder::new().num_threads(configuration.thread_count as usize).build_global().unwrap();
-
-	let code = if interative {
-		eprintln!("running in iteractive mode");
-
-		let app = App::configure(configuration);
-		app.run()
-	} else {
-		eprintln!("running in script mode");
-
-		let script = Script::configure(configuration);
-		script.run()
-	};
+	let launcher = Launcher::new();
+	let code = launcher.run();
 
 	exit(code);
 }